+++ src/lib.rs
@@ -9,5 +9,6 @@
 #![deny(dead_code)]
 #![deny(unused)]
 #![deny(clippy::allow_attributes_without_reason)]
+#![warn(clippy::missing_const_for_fn)]
 
 pub mod reindeer2;