+++ tests/system_testing/matrix_raw_merged/run.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-reindeer2="./target/debug/reindeer2"
-LOCAL_FOLDER="tests/system_testing/matrix_raw_merged"
-INPUT_FOF_0="$LOCAL_FOLDER/datasets/matrix_raw_0.fof"
-INPUT_FOF_1="$LOCAL_FOLDER/datasets/matrix_raw_1.fof"
-INPUT_FOF_INDEXES="$LOCAL_FOLDER/datasets/matrix_raw_index.fof"
-QUERY_INPUT="$LOCAL_FOLDER/query.fa"
-QUERY_OUTPUT="$LOCAL_FOLDER/results.tsv"
-EXPECTED_QUERY_OUTPUT="$LOCAL_FOLDER/expected.tsv"
+##### variable definitions #####
+reindeer2=$(realpath "./target/debug/reindeer2")  # we must define `reindeer2` as it is not installed yet
+LOCAL_FOLDER=$(dirname $0)
+INPUT_FOF_0="datasets/matrix_raw_0.fof"
+INPUT_FOF_1="datasets/matrix_raw_1.fof"
+INPUT_FOF_INDEXES="datasets/matrix_raw_index.fof"
+QUERY_INPUT="query.fa"
+QUERY_OUTPUT="results.tsv"
+EXPECTED_QUERY_OUTPUT="expected.tsv"
+
+cd $LOCAL_FOLDER
 
 cargo build --quiet
 RUST_LOG=warn $reindeer2 index --threads 7 --input $INPUT_FOF_0 -k 31 --nb-file-capacity 2 -o integration_test_index_0
@@ -15,7 +18,7 @@ RUST_LOG=warn $reindeer2 index --threads 7 --input $INPUT_FOF_1 -k 31 --nb-file-
 RUST_LOG=warn $reindeer2 merge --threads 7 --file-of-indexes $INPUT_FOF_INDEXES -o integration_test_index
 RUST_LOG=warn $reindeer2 query --fasta $QUERY_INPUT --index ./integration_test_index --output-format matrix-raw --output $QUERY_OUTPUT
 
-python3 "$LOCAL_FOLDER/files_equal.py" $QUERY_OUTPUT $EXPECTED_QUERY_OUTPUT
+python3 files_equal.py $QUERY_OUTPUT $EXPECTED_QUERY_OUTPUT
 is_same=$?
 
 rm -r ./integration_test_index_0